Added cost and move away from eventbrite
It seems as thoug we are not going to use eventbrite anymore so made the
verbage more genearic for just "regestration". Also added the ability to
choose between online and in person. Made both that and the course types
driven by a for loop and a global array of strings so new course types and
access types can be added easily. It also simplifies the form a bit.